Fix post-merge build breakage from origin/main merge
Reconciles callsites that referenced pre-merge field/global names against
their post-merge renamed counterparts, and restores four dropped includes
of ui_text_label_helpers_decls.h:

- TLoadSavePicture.cpp: g_szSaveFileReadBinaryMode_00698720 collided with
  the already-declared g_szLiteralRb_00698720 ("rb") at the same address;
  use the existing name.
- TMiniShipLine.cpp/.h: TMiniShipView::field84 was retyped to TShip*
  shipNode84 by main; retype TMiniShipLine::field10 to TShip* to match
  (it's assigned directly into shipNode84, no cast needed).
- TShipyardView.cpp/.h: fieldA0/fieldA2 were split/renamed to
  selectedRequirementRow (+0xa0) and an unnamed +0xa2 short by main;
  restore the fieldA2=0 write (real evidence from the original
  disassembly) by giving +0xa2 a scalar short type instead of leaving it
  an untouched byte array.
- TCreditsPicture.cpp, TQueryFloater.cpp, TSpecialQuitPicture.cpp: missing
  #include "game/ui_text_label_helpers_decls.h" for
  BuildUiTextStyleDescriptor/InitializeUiTextStyleDescriptor/
  MapUiThemeCodeToStyleFlags.

Verified: just build, just gates, just test all pass. just stats shows a
net improvement (+0.65pp aligned/original, +0.55pp avg similarity, 207
functions improved) with 24 minor score-wobble drops in functions this
change didn't touch (FP commutative-order/register-allocation wobble from
TU growth across this merge, per the documented score-wobble policy).
Baseline refreshed via just stats-baseline-update.

+++ src/game/TIndustryView.cpp
@@ -40,7 +40,7 @@ void TIndustryView::HandleEvent(int commandId, TEventHandler* sourceHandler, TEv
       FailNilPointerWithAssert("D:\\Ambit\\Cross\\UCityViews.cpp", 0x84c);
     }
 
-    g_pUiRuntimeContext->HandleTurnEventDialogFactorySlotB8(field9e, field94, (int)mainControl);
+    g_pUiRuntimeContext->HandleTurnEventDialogFactorySlotB8(field9e, (int)field94, (int)mainControl);
 
     TView* owner2 = OwnerPanel();
     static_cast<TWindow*>(owner2)->nativeWindow50->EnableWindow(wasDisabled);
